pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyntegritydb"
version = "0.1.0"
authors = [
  { name = "Osvaldo Trujillo", email = "osvaldo.trujillo.ingenieria@gmail.com" },
]
description = "Una biblioteca para medir la calidad de la integridad referencial en bases de datos relacionales."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
keywords = ["database", "integrity", "data quality", "sql", "validator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Database",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "psycopg2-binary==2.9.10",
    "sqlalchemy>=2.0",
    "networkx>=3.0",
    "pandas>=2.0",
    "tabulate>=0.9",
    "PyYAML>=6.0",
    "matplotlib>=3.7"
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "mock==5.2.0",
    "PyYAML>=6.0",
    "ruff"
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
]

[project.urls]
Homepage = "https://github.com/osvaldomx/pyntegritydb"
Repository = "https://github.com/osvaldomx/pyntegritydb"
Changelog = "https://github.com/osvaldomx/pyntegritydb/releases/tag/v0.2.0"
Issues = "https://github.com/osvaldomx/pyntegritydb/issues"

[project.scripts]
pyntegritydb = "pyntegritydb.cli:main"

[tool.pytest.ini_options]
norecursedirs = ["examples", "docs", "scripts"]